pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "community-crisis-predictor"
version = "0.1.0"
description = "Predict community mental health crises from Reddit data"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "praw>=7.7",
    "xgboost>=2.0",
    "scikit-learn>=1.3",
    "vaderSentiment>=3.3",
    "textstat>=0.7",
    "bertopic>=0.16",
    "sentence-transformers>=2.2",
    "shap>=0.43",
    "plotly>=5.18",
    "optuna>=3.4",
    "numpy>=1.24",
    "scipy>=1.11",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
